+++ src/page/visualization/nfaToDot.js
@@ -1,3 +1,8 @@
+
+function escapeAndPrettify(name) {
+    return name === " " ? "space" : name.replaceAll(/\\/g, "\\\\");
+}
+
 export default function nfaToDot(nfa) {
     let dot = `digraph nfa {
     rankdir=LR;
@@ -7,7 +12,7 @@ export default function nfaToDot(nfa) {
     "" -> ${nfa.initialState}\n`;
     for (const state of Object.values(nfa.states)) {
         for (const [matcher, toState] of state.transitions) {
-            dot += `    ${state.name} -> ${toState.name} [label = "${matcher.label}"]; \n`
+            dot += `    ${state.name} -> ${toState.name} [label = "${escapeAndPrettify(matcher.label)}"]; \n`
         }
     }
     return dot + "}";